Decompose backends/shared.py into cohesive modules
Split the 378-line dumping ground into five focused modules:
- labels.py: container label constants
- naming.py: resource naming and backend type helpers
- proxy_config.py: proxy/CA configuration, ProxyCredentials, credentials
- session_env.py: path encoding, session env building, sandbox config
- ssh.py: SSH backend construction

All 42+ import sites migrated to the new module paths. The old
shared.py is removed entirely (no re-export shim needed since all
consumers were updated in this change).
