+++ src/paude/cli/remote_git_setup.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 
 import typer
 
-from paude.backends.shared import (
+from paude.backends.naming import (
     engine_binary_for_backend,
     resource_name,
 )
