+++ tests/test_podman_proxy.py
@@ -13,7 +13,7 @@ from paude.backends.podman.proxy import (
     auth_volume_name,
     ca_volume_name,
 )
-from paude.backends.shared import derive_agent_ip
+from paude.backends.proxy_config import derive_agent_ip
 
 
 def _make_mock_runner(engine_binary: str = "podman") -> MagicMock:
@@ -471,7 +471,7 @@ class TestBuildCaBundleCmd:
     def test_cmd_writes_to_ca_bundle_path(self) -> None:
         """_BUILD_CA_BUNDLE_CMD must write to CA_BUNDLE_PATH."""
         from paude.backends.podman.proxy import _BUILD_CA_BUNDLE_CMD
-        from paude.backends.shared import CA_BUNDLE_PATH
+        from paude.backends.proxy_config import CA_BUNDLE_PATH
 
         assert CA_BUNDLE_PATH in _BUILD_CA_BUNDLE_CMD
