+++ tests/test_podman_session.py
@@ -8,6 +8,15 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from paude.backends.base import SessionConfig
+from paude.backends.labels import (
+    PAUDE_LABEL_AGENT,
+    PAUDE_LABEL_APP,
+    PAUDE_LABEL_CREATED,
+    PAUDE_LABEL_DOMAINS,
+    PAUDE_LABEL_PROXY_IMAGE,
+    PAUDE_LABEL_SESSION,
+    PAUDE_LABEL_WORKSPACE,
+)
 from paude.backends.podman import (
     PodmanBackend,
     SessionExistsError,
@@ -19,17 +28,8 @@ from paude.backends.podman.helpers import (
     find_container_by_session_name,
 )
 from paude.backends.podman.session_setup import SessionSetup
-from paude.backends.shared import (
-    PAUDE_LABEL_AGENT,
-    PAUDE_LABEL_APP,
-    PAUDE_LABEL_CREATED,
-    PAUDE_LABEL_DOMAINS,
-    PAUDE_LABEL_PROXY_IMAGE,
-    PAUDE_LABEL_SESSION,
-    PAUDE_LABEL_WORKSPACE,
-)
-from paude.backends.shared import decode_path as _decode_path_raw
-from paude.backends.shared import encode_path as _encode_path_raw
+from paude.backends.session_env import decode_path as _decode_path_raw
+from paude.backends.session_env import encode_path as _encode_path_raw
 
 
 def encode_path(path: Path) -> str:
